precio como un entero

+++ SistemaTurnos.WebApplication/WebApi/Dto/Subspecialty/HairdressingAddSubspecialtyDto.cs
@@ -2,6 +2,6 @@
 {
     public class HairdressingAddSubspecialtyDto : AddSubspecialtyDto
     {
-        public decimal Price { get; set; }
+        public uint Price { get; set; }
     }
 }
